@@ -1,18 +1,19 @@
 [project]
 name = "nonebot-plugin-miragetank"
-version = "0.1.5"
+version = "0.2.0"
 description = "生成幻影坦克图，在不同背景下显示不同图片"
 authors = [
     {name = "1umine", email = "a1628420979@163.com"},
 ]
 dependencies = [
-    "numpy>1.24,<2.0",
-    "aiohttp>=3.8.0,<4.0",
-    "Pillow>=9.0.0,<10.0",
+    "numpy>=1.24,<2.0",
+    "Pillow>=9.0.0",
     "nonebot2>=2.0,<3.0",
-    "nonebot-adapter-onebot>=2.0,<3.0",
+    "httpx>=0.27.0",
+    "nonebot-plugin-alconna>=0.42.4",
+    "nonebot-plugin-waiter>=0.6.2",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 readme = "README.md"
 license = {text = "GPL"}
 
